requirements.txt
Flask
stripe
pyyaml